pyproject.toml
[tool.poetry]
name = "reactome_to_bel"
version = "0.1.0"
description = "Convert reactome to BEL"
authors = ["William Hayes <william.s.hayes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
requests = "*"
structlog = "*"
python-arango = "*"
invoke = "*"
jsons = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
flake8 = "*"
pylint = "*"
pep8 = "*"
logging-tree = "*"
devtools = {extras = ["pygments"],version = "*"}


[tool.black]
line-length = 100
target-version = ['py37', 'py38']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  # The following are specific to Black, you probably don't want those.
  | blib2to3
  | tests/data
  | profiling
)/
'''
[tool.isort]
multi_line_output=3
include_trailing_comma=true
force_grid_wrap=false
use_parentheses=true
line_length=100
import_heading_stdlib="Standard Library"
import_heading_thirdparty="Third Party Imports"
import_heading_firstparty="Local Imports"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


